krishi_sahayak/requirements.txt
# Web Framework and API
Flask==2.3.3
gunicorn==21.2.0

# Scraping and PDF Processing
beautifulsoup4==4.12.2
scrapy==2.11.0
PyMuPDF==1.23.7
pytesseract==0.3.10
google-cloud-vision==3.4.4

# NLP and ML
transformers==4.36.2
torch==2.2.0
sentencepiece==0.1.99
langdetect==1.0.9
fasttext==0.9.2
indicnlp==0.1
googletrans==3.1.0a0
openai-whisper==20231117

# WhatsApp Integration
twilio==8.10.3

# Database and Message Queue
pymongo==4.6.1
kafka-python==2.0.2

# Cloud and Deployment
boto3==1.34.14
google-cloud-scheduler==2.12.0
kubernetes==28.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
pillow==10.2.0
pytest==7.4.3
logging==0.4.9.6 